+++ healthwatchfeedback.php
@@ -10,9 +10,9 @@ defined( 'ABSPATH' ) or die( 'Sorry, nothing to see here.' );
 
 // Includes
 
-include('post-types-taxononmies.php'); 
+include('post-types-taxononmies.php');
 
-include('taxonomy-icons.php'); 
+include('taxonomy-icons.php');
 
 include('feedback-form.php');
 
@@ -29,14 +29,11 @@ include('widgets/widget-enter-and-view.php');
 
 include('widgets/widget-recent-comments.php');
 
-// Include shortcodes
-
-include('shortcodes/shortcode-ratings-block.php');
-
 // Add functions
 
 require_once('functions/functions-star-rating.php');
 require_once('functions/functions-get-rating.php');
+require_once('functions/shortcodes.php');
 
 // Enqueue CSS
 
@@ -44,19 +41,6 @@ require_once('functions/functions-get-rating.php');
     function hw_enqueue_css() {
         wp_enqueue_style( 'prefix-style', plugins_url('css/style.css', __FILE__) );
     }
-
-
-// Form validation
-
-function load_form_validation_js()
-{
-    // Register the script like this for a plugin:
-    wp_register_script( 'form-validation', plugins_url( '/javascript/form-validation.js', __FILE__ ) );
- 
-    // For either a plugin or a theme, you can then enqueue the script:
-    wp_enqueue_script( 'form-validation' );
-}
-add_action( 'wp_enqueue_scripts', 'load_form_validation_js' );
 
 // add fontawesome on edit-comments.php
 function add_fontawesome_edit_comments( $hook ) {